Safari-611.1.10/LayoutTests/fast/dom/xml-parser-entity-in-attribute-value-expected.txt

This test verifies the parsing of unicode html entities used as an invalid style attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS paragraphElement.getAttribute("style") is "0񶔬"
PASS paragraphElement.style.cssText is ""
PASS successfullyParsed is true

TEST COMPLETE
x
